Move game renderer mixin into impl subpackage; Example code cleanup
