+++ src/main/java/de/florianmichael/imguiexample/mixin/TitleScreenMixin.java
@@ -8,16 +8,15 @@ import org.spongepowered.asm.mixin.Mixin;
 
 @Mixin(TitleScreen.class)
 public class TitleScreenMixin implements RenderInterface {
+
     @Override
     public void render(ImGuiIO io) {
-        // Example on how to use a custom Font
-        // ImGui.pushFont(ImGuiImpl.defaultFont);
         ImGui.begin("Hello World");
         // Draw something here, see the official example module for more information:
         // https://github.com/ocornut/imgui/blob/master/imgui_demo.cpp
         ImGui.end();
 
         ImGui.showDemoWindow();
-        // ImGui.popFont();
     }
+
 }
