+++ src/main/java/de/florianmichael/imguiexample/screens/ExampleScreen.java
@@ -3,12 +3,14 @@ package de.florianmichael.imguiexample.screens;
 import de.florianmichael.imguiexample.imgui.RenderInterface;
 import imgui.ImGui;
 import imgui.ImGuiIO;
-import imgui.flag.ImGuiWindowFlags;
 import imgui.type.ImBoolean;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.text.Text;
 
 public final class ExampleScreen extends Screen implements RenderInterface {
+
+    private static final ImBoolean showDemoWindow = new ImBoolean(false);
+
     public ExampleScreen() {
         super(Text.literal("Example Screen"));
     }
@@ -18,18 +20,15 @@ public final class ExampleScreen extends Screen implements RenderInterface {
         return false; // Only relevant in singleplayer
     }
 
-
-    private static final ImBoolean FALSE = new ImBoolean(false);
-
     @Override
     public void render(ImGuiIO io) {
-        if (ImGui.begin("Hello, World!", ImGuiWindowFlags.None)) {
+        if (ImGui.begin("Hello, World!")) {
             ImGui.setWindowSize(800, 600);
-
-            ImGui.checkbox("Show Demo Window", FALSE);
+            ImGui.checkbox("Show Demo Window", showDemoWindow);
+            ImGui.end();
         }
-        ImGui.end();
 
-        ImGui.showDemoWindow(FALSE);
+        ImGui.showDemoWindow(showDemoWindow);
     }
+
 }
